+++ lib/keeper/keeper_heartbeat_loop.ml
@@ -383,7 +383,7 @@ let run_keepalive_unified_turn
             ~runtime_profile:(runtime_id_of_meta meta_after_triage)
             ~keeper_name:meta_after_triage.name
             ~channel:turn_decision.channel
-            (fun ~semaphore_wait_ms ->
+            (fun ~semaphore_wait_ms ~slot_control ->
                run_keeper_cycle_with_slot
                  ~ctx
                  ~meta_after_cursor_persist
@@ -392,6 +392,7 @@ let run_keepalive_unified_turn
                  ~turn_decision
                  ~shared_context
                  ~semaphore_wait_ms
+                 ~slot_control
                  ())
         with
         | Ok meta -> meta