+++ lib/keeper/keeper_tools_oas_bundle.ml
@@ -32,6 +32,7 @@ let make_tool_bundle
       ?search_fn
       ?on_tool_called
       ?clock
+      ?turn_slot_control
       ()
   : tool_bundle
   =
@@ -113,6 +114,7 @@ let make_tool_bundle
                ?search_fn
                ?on_tool_called
                ?clock
+               ?turn_slot_control
                ~failure_counts
                ()
            in
@@ -167,6 +169,7 @@ let make_tool_bundle
                ?search_fn
                ?on_tool_called
                ?clock
+               ?turn_slot_control
                ~translate_input:descriptor.translate
                ~failure_counts
                ()
@@ -200,9 +203,18 @@ let make_tools
       ?search_fn
       ?on_tool_called
       ?clock
+      ?turn_slot_control
       ()
   : Agent_sdk.Tool.t list
   =
-  (make_tool_bundle ~config ~meta ~ctx_snapshot ?search_fn ?on_tool_called ?clock ())
+  (make_tool_bundle
+     ~config
+     ~meta
+     ~ctx_snapshot
+     ?search_fn
+     ?on_tool_called
+     ?clock
+     ?turn_slot_control
+     ())
     .tools
 ;;